Share ZooKeeper conn. between locks
There seems to be an upper limit to the number of ZooKeeper connections
that a process can have and that does make sense ($N heartbeats? ouch).
Thus, we share the zookeeper connection between locks and release it
whenever the last lock using the connection is released.

For some reason, Net::ZooKeeper::Lock will slow down tremendously with
the number of procs in the same lock set. This must be related to the
number of sequence-numbered znodes or something like that. TBD.